fix glob error for some OSes

+++ src/FileCache.php
@@ -190,12 +190,14 @@ class FileCache implements CacheInterface {
 			}
 		}
 		$files = glob($this->getDirectory()."*");
-        $prefixFile = str_replace(array("_", "/", "\\", ":"), array("___", "_s_", "_b_", "_d_"), $this->prefix);
-		foreach ($files as $filename) {
-			if (empty($prefixFile) || strpos(basename($filename), $prefixFile) === 0) {
-		    	unlink($filename);
-			}
-		}
+        if ($files !== false){// some file systems wont distinguish between empty match and an error
+            $prefixFile = str_replace(array("_", "/", "\\", ":"), array("___", "_s_", "_b_", "_d_"), $this->prefix);
+            foreach ($files as $filename) {
+                if (empty($prefixFile) || strpos(basename($filename), $prefixFile) === 0) {
+                    unlink($filename);
+                }
+            }
+        }
 	}
 	
 	protected function getDirectory() {
